refactor: move printConfirmSubscription to middleware
