+++ backend/src/email/middlewares/email-callback.middleware.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express'
 import { EmailCallbackService } from '@email/services'
+import logger from '@core/logger'
 
 const isAuthenticated = (
   req: Request,
@@ -30,7 +31,28 @@ const parseEvent = async (
   }
 }
 
+const printConfirmSubscription = (
+  req: Request,
+  res: Response,
+  next: NextFunction
+): Response | void => {
+  const { Type: type, SubscribeURL: subscribeUrl } = req.body
+  if (type === 'SubscriptionConfirmation') {
+    const parsed = new URL(req.body['SubscribeURL'])
+    if (
+      parsed.protocol === 'https:' &&
+      /^sns\.[a-zA-Z0-9-]{3,}\.amazonaws\.com(\.cn)?$/.test(parsed.host)
+    ) {
+      logger.info(
+        `To confirm the subscription, enter this url: ${subscribeUrl}`
+      )
+      return res.sendStatus(202)
+    }
+  }
+  return next()
+}
 export const EmailCallbackMiddleware = {
   isAuthenticated,
   parseEvent,
+  printConfirmSubscription,
 }
